requirements.txt
# Add your project dependencies here
colorgram.py==1.2.0